Add algorithms_map and benchmark_scheduler modules for sorting algorithm execution and iteration management
