+++ algorithms_map.py
@@ -1,0 +1,57 @@
+# algorithms_map.py
+from algorithms import *  # Import your individual sorting functions
+
+def get_algorithms():
+    return {
+        "Bead Sort": bead_sort,
+        "Bitonic Sort Parallel": bitonic_sort_parallel,
+        "Block Sort": block_sort,
+        "Bogo Sort": bogo_sort,
+        "Bubble Sort": bubble_sort,
+        "Bucket Sort": bucket_sort,
+        "Burst Sort": burst_sort,
+        "Cocktail Sort": cocktail_sort,
+        "Comb Sort": comb_sort,
+        "Counting Sort": counting_sort,
+        "Cubesort": cubesort,
+        "Cycle Sort": cycle_sort,
+        "Exchange Sort": exchange_sort,
+        "Flash Sort": flash_sort,
+        "Franceschini's Method": franceschinis_method,
+        "Gnome Sort": gnome_sort,
+        "Heap Sort": heap_sort,
+        "Hyper Quick": hyper_quick,
+        "I Can't Believe It Can Sort": i_cant_believe_it_can_sort,
+        "Insertion Sort": insertion_sort,
+        "Intro Sort": intro_sort,
+        "Library Sort": library_sort,
+        "LSD Radix Sort": lsd_radix_sort,
+        "Merge Insertion Sort": merge_insertion_sort,
+        "Merge Sort": merge_sort,
+        "Merge Sort In-Place": merge_sort_inplace,
+        "MSD Radix Sort": msd_radix_sort,
+        "MSD Radix Sort In-Place": msd_radix_sort_inplace,
+        "Odd-Even Sort": odd_even_sort,
+        "Pancake Sort": pancake_sort,
+        "Patience Sort": patience_sort,
+        "Pigeonhole Sort": pigeonhole_sort,
+        "Polyphase Merge Sort": polyphase_merge_sort,
+        "Postman Sort": postman_sort,
+        "Quick Sort": quick_sort,
+        "Radix Sort": radix_sort,
+        "Replacement Selection Sort": replacement_selection_sort,
+        "Sample Sort": sample_sort,
+        "Selection Sort": selection_sort,
+        "Shell Sort": shell_sort,
+        "Sleep Sort": sleep_sort,
+        "Slowsort": slowsort,
+        "Smooth Sort": smooth_sort,
+        "Sorting Network": sorting_network,
+        "Spaghetti Sort": spaghetti_sort,
+        "Spreadsort": spreadsort,
+        "Stooge Sort": stooge_sort,
+        "Strand Sort": strand_sort,
+        "Tim Sort": tim_sort,
+        "Tournament Sort": tournament_sort,
+        "Tree Sort": tree_sort,
+    }
